requirements.txt
python-documentcloud==3.0.0
requests
